backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
requests
numpy
python-multipart
